Refactor BalanceTable and OrderTable to favor performance
